// interfaces/bean-info-pattern.ts
// Generated by https://quicktype.io

export interface TopLevel {
    DataBeanProperties: TopLevelDataBeanProperties;
    DataBeanName: string;
}

export interface TopLevelDataBeanProperties {
    Type: number;
    ObjectValue: ObjectValue;
    ClassName: string;
    Date: string;
}

export interface ObjectValue {
    DataBeanProperties: BeanInfoPattern;
    DataBeanName: string;
}

export interface BeanInfoPattern {
    ClassForName: string;
    SequenceName: string | null;
    RemoteLinkType: null;
    Description: string | null;
    IDDaoInfo: number | null;
    UseRemoteLinkForSequence: null;
    URL: null;
    Name: string | null;
    IDDataBeanInfo: number;
    HashCode: null;
    RemoteLinkPropertyName: null;
    State: null;
    DataBeanTable: string;
    Since: string;
    UseRemoteLinkForTable: null;
}