// interfaces/bean-properties.ts
// Generated by https://quicktype.io

export interface BeanProperty {
    IDDataBeanProperty: number;
    ClassForName: string;
    Description: null;
    DBForeignColumnName: null;
    DBDataType: null;
    OriginDataBean: null;
    MigrationType: null;
    DBType: null;
    Name: string;
    IsCreated: null;
    Since: string;
    DBLength: null;
    IDDBTable: null;
    OriginProperty: null;
    IDDataType: null;
    IsMediaFile: null;
    PropertyIsNotNull: null;
    DBIDColumn: null;
    IDForeignKey: null;
    TableKey: boolean;
    DBForeignKeyName: null;
    DBScale: null;
    IDDataBeanInfo: number;
    PropertyDefaultValue: null;
    State: number;
    DBForeignTableName: null;
}